refactor(Market): extract MarketOutcome component
